Krishnamoorthy number check
A programme to check whether the given number is a krishnamoorthy number or not. Krishnamoorthy number is a number in which the sum of the factorial of the digits of a number will equal the number for eg 145 is a krishnamoorthy number. Interestingly 1, 2, 145, and 40585 are the only four krishnamoorthy numbers known to us.